+++ functions.php
@@ -6,17 +6,14 @@ use Exception;
 define(__NAMESPACE__ .'\THEME', 'sim-theme');
 define(__NAMESPACE__ .'\THEME_PATH', str_replace('\\', '/', __DIR__));
 
-//check if plugin is already installed
-$activePlugins	= get_option( 'active_plugins' );
-
 // We cannot use the sim theme
-if(!in_array('sim-base/sim-base.php', $activePlugins) && !in_array('sim-plugin/sim-plugin.php', $activePlugins)){
+if(is_plugin_inactive('tsjippy-shared-functionality/tsjippy-shared-functionality.php') && is_plugin_inactive('sim-plugin/sim-plugin.php')){
     $themes = wp_get_themes();
     unset($themes['SIM theme']);
 
     switch_theme(array_values($themes)[0]['theme_root']);
-    error_log("To use the sim-theme you need to install the sim-base");
-    //throw new Exception("To use the sim-theme you need to install the sim-base");
+    error_log("To use the sim-theme you need to install the tsjippy-shared-functionality plugin");
+    //throw new Exception("To use the sim-theme you need to install the tsjippy-shared-functionality");
 }
 
 // composer